// packages/extension/src/types/provider.ts
export enum ProviderName {
  samara = "samara",
  radiant = "radiant",
}
// ...
export interface AssetsType {
  name: string;
  symbol: string;
  icon: string;
  balance: string;
  balancef: string;
  balanceUSD: number;
  balanceUSDf: string;
  value: string;
  valuef: string;
  contract?: string;
  decimals: number;
  sparkline: string;
  priceChangePercentage: number;
  baseToken?: BaseToken;
  unspentCoins: UTXO[];
}